app.requirements.txt
flask
pydantic
pydantic[email]
uvicorn
black
torch
torchvision
pillow
albumentations
python-box